requirements.txt
pandas
openpyxl
zipfile
base64
folium
streamlit_folium
matplotlib
utm
numpy
scipy
geojsoncontour
